Added a single REA base class for all the polymorphic REA models to inherit from. I think they should all be polymorphic to enable things like searching for an object from just a uuid.
Added a couple of quick tests to confirm that the polymorphism is
working the way that I was hoping. Seems to be okay.

Added the migration back in that seemed to go missing at some point.
Creating the migration is a little bit annoying, since you need a
django project to do it (I think).

Updated setup.py and requirements.txt to include django-uuidfield
from the always rad dcramer.